+++ CompactDS/Bitmaps/DiffSetRL2_64.cs
@@ -37,13 +37,13 @@ namespace natix.CompactDS
 		protected int M;
 		protected short B = 31;
 		IList<long> Samples;
-		IList<int> Offsets;
+		IList<long> Offsets;
 		// int run_len = 0;
 		
 		public DiffSetRL2_64 ()
 		{
 			this.Samples = new List<long> ();
-			this.Offsets = new List<int> ();
+			this.Offsets = new List<long> ();
 			this.Stream = new BitStream32 ();
 			this.Coder = new EliasDelta64 ();
 		}
@@ -88,7 +88,7 @@ namespace natix.CompactDS
 				}
 			} else {*/
 				PrimitiveIO<long>.WriteVector (Output, this.Samples);
-				PrimitiveIO<int>.WriteVector (Output, this.Offsets);
+				PrimitiveIO<long>.WriteVector (Output, this.Offsets);
 			//}
 			this.Stream.Save (Output);
 		}
@@ -114,9 +114,9 @@ namespace natix.CompactDS
 				}
 			} else {*/
 				this.Samples = new long[ num_samples ];
-				this.Offsets = new int[ num_samples ];
+				this.Offsets = new long[ num_samples ];
 				PrimitiveIO<long>.ReadFromFile (Input, num_samples, this.Samples);
-				PrimitiveIO<int>.ReadFromFile (Input, num_samples, this.Offsets);
+				PrimitiveIO<long>.ReadFromFile (Input, num_samples, this.Offsets);
 			//}
 			// POS = R.BaseStream.Position - POS;
 			// Console.WriteLine("=======*******=======>> POS: {0}", POS);
@@ -140,7 +140,7 @@ namespace natix.CompactDS
 				throw new ArgumentException ("DiffSetRL2_64 B difference");
 			}
 			Assertions.AssertIList<long> (this.Samples, other.Samples, "DiffSetRL2_64 Samples difference");
-			Assertions.AssertIList<int> (this.Offsets, other.Offsets, "DiffSetRL2_64 Offsets difference");
+			Assertions.AssertIList<long> (this.Offsets, other.Offsets, "DiffSetRL2_64 Offsets difference");
 			this.Stream.AssertEquality (other.Stream);
 		}
 		
@@ -363,7 +363,7 @@ namespace natix.CompactDS
 				if ((this.M % this.B) == 0) {	
 					this.Commit (ctx);
 					this.Samples.Add (current);
-					this.Offsets.Add ((int)this.Stream.CountBits);
+					this.Offsets.Add (this.Stream.CountBits);
 				}
 				if (current >= this.N) {
 					this.N = current + 1;
